Implements landing page for Marc et Cécile
Creates the initial landing page for the Marc et Cécile website.

This includes adding images, content sections for the boutique,
team information, and social media links. Also includes a section with a map.

Removes unused images, renames an image for clarity, and introduces
new images to enhance the visual appeal of the landing page.